Push down javadoc code from jdt.ui to jdt.core.manipulations

WIP - push down and break up logic for javadoc html generation


Comment out unused methods


Upversion some stuff


Upversion some stuff


Bad since tag


Compilation error


Revert "TEMPORARY patch to allow access to expected org.eclipse.text classes. Revert before merge"

This reverts commit ebe0ea85973a5ee40e0c0752adf4149ba6bf455a.

TEMPORARY patch to allow access to expected org.eclipse.text classes. Revert before merge


JDT null check

Bugfix for test suite


Some cleanup, delete commented out implementations, etc


Fix a jdt.ls test case, revert if breaks jdt


Continue work for abstraction points for jdt.ls, copy temporary changes to org.eclipse.text packages


Revert "Work around API tooling warning regarding unnecessary minor bump."

This reverts commit 19a3a4874a81df63625acfc5d6969b23ff9fdccc.

Make changes to temporary files to get a green build


Remove temporary workarounds and clean up code


Organize whitespace, remove unused strings


viewsupport package needs split package flag
