+++ org.eclipse.jdt.core.manipulation/.settings/.api_filters
@@ -1,13 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <component id="org.eclipse.jdt.core.manipulation" version="2">
-    <resource path="META-INF/MANIFEST.MF">
-        <filter comment="See https://github.com/eclipse-jdt/eclipse.jdt.ui/pull/914" id="926941240">
-            <message_arguments>
-                <message_argument value="1.21.0"/>
-                <message_argument value="1.20.0"/>
-            </message_arguments>
-        </filter>
-    </resource>
     <resource path="common/org/eclipse/jdt/core/manipulation/ImportReferencesCollector.java" type="org.eclipse.jdt.core.manipulation.ImportReferencesCollector">
         <filter id="576720909">
             <message_arguments>
