+++ org.eclipse.jdt.core.manipulation/META-INF/MANIFEST.MF
@@ -36,6 +36,7 @@ Export-Package: org.eclipse.jdt.core.manipulation,
  org.eclipse.jdt.internal.corext.dom.fragments;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
  org.eclipse.jdt.internal.corext.fix;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
  org.eclipse.jdt.internal.corext.fix.helper;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
+ org.eclipse.jdt.core.manipulation.internal.javadoc;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
  org.eclipse.jdt.internal.corext.refactoring;manipulation=split;mandatory:=manipulation;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
  org.eclipse.jdt.internal.corext.refactoring.base;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
  org.eclipse.jdt.internal.corext.refactoring.binary;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
@@ -79,6 +80,7 @@ Export-Package: org.eclipse.jdt.core.manipulation,
  org.eclipse.jdt.internal.ui.text.correction.proposals;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
  org.eclipse.jdt.internal.ui.text.template.contentassist;x-friends:="org.eclipse.jdt.ui, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
  org.eclipse.jdt.internal.ui.util;x-friends:="org.eclipse.jdt.ui,org.eclipse.jdt.astview,org.eclipse.jdt.junit.core, org.eclipse.jdt.ui.tests, org.eclipse.jdt.ui.tests.refactoring, org.eclipse.jdt.text.tests",
+ org.eclipse.jdt.internal.ui.viewsupport;manipulation=split;mandatory:=manipulation,
  org.eclipse.jdt.ui.refactoring;manipulation=split;mandatory:=manipulation,
  org.eclipse.jdt.ui.text.java.correction;manipulation=split;mandatory:=manipulation;x-friends:="org.eclipse.jdt.ui"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
