+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/text/javadoc/JavaDocMessages.properties
@@ -15,21 +15,6 @@
 JavaDoc2HTMLTextReader_parameters_section=Parameters:
 JavaDoc2HTMLTextReader_returns_section=Returns:
 JavaDoc2HTMLTextReader_throws_section=Throws:
-JavaDoc2HTMLTextReader_type_parameters_section=Type Parameters:
 JavaDoc2HTMLTextReader_author_section=Author:
-JavaDoc2HTMLTextReader_deprecated_section=Deprecated.
-JavaDoc2HTMLTextReader_method_in_type={0} in {1}
-JavaDoc2HTMLTextReader_overrides_section=Overrides:
 JavaDoc2HTMLTextReader_see_section=See Also:
 JavaDoc2HTMLTextReader_since_section=Since:
-JavaDoc2HTMLTextReader_specified_by_section=Specified by:
-JavaDoc2HTMLTextReader_version_section=Version:
-JavaDoc2HTMLTextReader_api_note=API Note:
-JavaDoc2HTMLTextReader_impl_spec=Impl Spec:
-JavaDoc2HTMLTextReader_impl_note=Impl Note:
-JavaDoc2HTMLTextReader_uses=Uses:
-JavaDoc2HTMLTextReader_provides=Provides:
-JavadocContentAccess2_getproperty_message=<p>Gets the value of the property {0}.</p><dl><dt>Property Description:</dt><dd>{1}</dd></dl>
-JavadocContentAccess2_setproperty_message=<p>Sets the value of the property {0}.</p><dl><dt>Property Description:</dt><dd>{1}</dd></dl>
-JavadocContentAccess2_returns_pre=Returns 
-JavadocContentAccess2_returns_post=.
